+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 ## Added
 * Attempt to use `psutil` (if available) to find interface MACs on all platforms
 * Attempt to use `netifacts` (if available) to find interface MAC on Non-Windows platforms
+
 ## Changed
-*
+* Significantly improved the detection of the default interface
 
 ## Fixed
 *